+++ Cargo.toml
@@ -42,7 +42,7 @@ image = { version = "0.24", default-features = false, features = ["png", "jpeg",
 indexmap = "1.9.3"
 log = "0.4"
 miniz_oxide = "0.7"
-oklab = "1.0.0"
+oklab = "1"
 once_cell = "1"
 pdf-writer = "0.7.1"
 pixglyph = "0.1"
